Renamed Tactics to UnitsControl. While Tactics is cooler, UnitsControl is easier to understand.
