+++ Bot/Managers/WarManagement/ArmySupervision/UnitsControl/SneakAttack/TerminalState.cs
@@ -1,8 +1,8 @@
 ﻿using System.Collections.Generic;
 
-namespace Bot.Managers.WarManagement.ArmySupervision.Tactics.SneakAttack;
+namespace Bot.Managers.WarManagement.ArmySupervision.UnitsControl.SneakAttack;
 
-public partial class SneakAttackTactic {
+public partial class SneakAttackUnitsControl {
     public class TerminalState: SneakAttackState {
         public override bool IsViable(IReadOnlyCollection<Unit> army) {
             return false;
