+++ Bot/VideoClips/VideoClipPlayer.cs
@@ -7,7 +7,6 @@ using Bot.Debugging;
 using Bot.GameData;
 using Bot.GameSense;
 using Bot.Utils;
-using Bot.VideoClips.Clips;
 using Bot.VideoClips.Clips.RayCastingClips;
 using Bot.Wrapper;
 using SC2APIProtocol;
